Moved adapters into the codetocad folder
