+++ codetocad/adapters/blender/blender_actions/modifiers.py
@@ -1,7 +1,7 @@
 from codetocad.core.angle import Angle
 from codetocad.core.dimension import Dimension
 from codetocad.enums.axis import Axis
-from adapters.blender.blender_definitions import (
+from codetocad.adapters.blender.blender_definitions import (
     BlenderBooleanTypes,
     BlenderLength,
     BlenderModifiers,
